pyproject.toml
[project]
name = "2026-practice-frameworks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain[anthropic]>=1.2.3",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
